// src/common/constants/about.enum.ts
export enum AboutCategoryType {
  TAB = 'tab',
  MENU = 'menu',
}

export enum MenuKey {
  EXPERIENCE = 'experience',
  HARD_SKILLS = 'hard-skills',
  SOFT_SKILLS = 'soft-skills',
  BIO = 'bio',
  INTERESTS = 'interests',
  EDUCATION = 'education',
  SPORTS = 'sports',
  FAVORITE_GAMES = 'favorite-games',
  FAVORITE_FOODS = 'favorite-foods',
}